// src/types/enums/Routes.ts
export enum EInternalRoute {
  Home = '/',
  Resume = '/resume',
  About = '/about'
}

export enum EExternalRoute {
  Gedcom = '/git/gedcom',
  Zillow = '/git/zillow',
  JimApp = '/git/jim',
  LiveGithub = '/git/github'
}

export enum EGithubRepoLink {
  Gedcom = 'https://github.com/ming1in/gedcom-parser',
  Zillow = 'https://github.com/ming1in/Stevens-IT/tree/master/CS_559/projects/03_zillow',
  JimApp = 'https://github.com/ming1in/jim-app',
  LiveGithub = 'https://github.com/WellSheet/live-github',
  Plaf = 'https://github.com/ming1in/Stevens-IT/tree/master/CS_496/interpreter'
}